Fix broken test configurations
