+++ .Package.test.swift
@@ -10,6 +10,7 @@ let package = Package(
     ],
     exclude: [
       "Sources/QuickObjectiveC",
+      "Tests/QuickTests/QuickAfterSuiteTests/AfterSuiteTests+ObjC.m",
       "Tests/QuickTests/QuickFocusedTests/FocusedTests+ObjC.m",
       "Tests/QuickTests/QuickTests/FunctionalTests/ObjC",
       "Tests/QuickTests/QuickTests/Helpers",
