--- Sources/QuickAfterSuiteTests/AfterSuiteTestsHeader.h
@@ -1,17 +1,0 @@
-//
-//  AfterSuiteTestHeader.h
-//  Quick
-//
-//  Created by Fujiki Takeshi on 2016/08/24.
-//  Copyright © 2016年 Brian Ivan Gesiak. All rights reserved.
-//
-
-#ifndef AfterSuiteTestHeader_h
-#define AfterSuiteTestHeader_h
-
-static BOOL afterSuiteFirstTestExecuted = NO;
-
-static BOOL afterSuiteTestsWasExecuted_ObjC = NO;
-static BOOL afterSuiteTestsWasExecuted = NO;
-
-#endif /* AfterSuiteTestHeader_h */
