+++ Tests/QuickTests/QuickAfterSuiteTests/Info.plist
