+++ Tests/QuickTests/QuickTests/Helpers/QuickTestsBridgingHeader.h
@@ -1,3 +1,2 @@
 #import "QCKSpecRunner.h"
 #import "QuickSpec+QuickSpec_MethodList.h"
-#import "AfterSuiteTestsHeader.h"
